Rename DoubleColumnRows to DoubleColumnGrid.
