+++ services/app-web/src/components/DoubleColumnGrid/index.ts
@@ -1,0 +1,1 @@
+export { DoubleColumnGrid } from './DoubleColumnGrid'
