--- services/app-web/src/components/DoubleColumnRows/index.ts
@@ -1,1 +1,0 @@
-export { DoubleColumnRows } from './DoubleColumnRows'
